requirements.txt
fastapi==0.115.6
python-multipart==0.0.9
pydantic==2.10.4
groq==0.13.0
pypdf==5.1.0
python-docx==1.1.2
reportlab==4.2.5
